+++ doc/change_log.md
@@ -29,7 +29,9 @@
 * Improving face-vertex accuracy when transferring mesh attributes
 * Replacing ConvexBlockData with MeshVectorData which provides a flexible way to add data layers to mesh features
 * Now including .import files when creating addon archive
-* Creating action to export Cyclops scenes in a custom file formst
+* Creating action to export Cyclops scenes in a custom file format
+* Adding new menubar for Cyclops that will always be shown in the toolbar
+* Adding Import MeshInstance command